fix: extract ElevationProfileView + SummaryCard to fix file_length lint
WalkSummaryView.swift was 1089 lines (limit: 1000). Extracted
ElevationProfileView and SummaryCard into their own file, bringing
WalkSummaryView to 990 lines. Also fix trailing newline in AppDelegate.
